Unnecessary casts
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/math/trunk@1150381 13f79535-47bb-0310-9956-ffa450edef68

+++ src/test/java/org/apache/commons/math/geometry/partitioning/utilities/AVLTreeTest.java
@@ -157,7 +157,7 @@ public class AVLTreeTest {
     }
 
     private int value(AVLTree<Integer>.Node node) {
-        return ((Integer) node.getElement()).intValue();
+        return node.getElement().intValue();
     }
 
     private void checkOrder(AVLTree<Integer> tree) {
